# services/server/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hot Cross Buns Sync Server API",
    "description": "REST API for the Hot Cross Buns sync server. All endpoints except /health and /api/v1/auth/* require a Bearer JWT token when AUTH_REQUIRED=true.",
    "version": "1.0.0",
    "contact": {
      "name": "Hot Cross Buns"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256 JWT with sub=userID, exp=30 days. Not required when AUTH_REQUIRED=false."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "One or more fields are invalid."
              },
              "details": {
                "type": "array",
                "items": { "type": "string" },
                "example": []
              }
            },
            "required": ["code", "message", "details"]
          }
        },
        "required": ["error"]
      },
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "urgent" },
          "color": { "type": "string", "nullable": true, "example": "#f38ba8" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "userId", "name", "createdAt"]
      },
      "CreateTagRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 255, "example": "urgent" },
          "color": { "type": "string", "nullable": true, "example": "#f38ba8" }
        },
        "required": ["name"]
      },
      "UpdateTagRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 255 },
          "color": { "type": "string", "nullable": true }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "listId": { "type": "string", "format": "uuid" },
          "parentTaskId": { "type": "string", "format": "uuid", "nullable": true },
          "title": { "type": "string", "example": "Buy groceries" },
          "content": { "type": "string", "nullable": true },
          "priority": { "type": "integer", "minimum": 0, "maximum": 3, "description": "0=none, 1=low, 2=medium, 3=high" },
          "status": { "type": "integer", "minimum": 0, "maximum": 1, "description": "0=open, 1=completed" },
          "dueDate": { "type": "string", "format": "date-time", "nullable": true },
          "dueTimezone": { "type": "string", "nullable": true, "example": "Asia/Singapore" },
          "recurrenceRule": { "type": "string", "nullable": true, "description": "RFC 5545 RRULE string" },
          "sortOrder": { "type": "integer" },
          "completedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "deletedAt": { "type": "string", "format": "date-time", "nullable": true },
          "subtasks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Task" }
          },
          "tags": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Tag" }
          }
        },
        "required": ["id", "userId", "listId", "title", "priority", "status", "sortOrder", "createdAt", "updatedAt"]
      },
...
      "ChangeRecord": {
        "type": "object",
        "properties": {
          "entityType": { "type": "string", "enum": ["list", "task", "tag"], "example": "task" },
          "entityId": { "type": "string", "format": "uuid" },
          "fieldName": { "type": "string", "example": "title" },
          "newValue": { "description": "Raw JSON value (string, number, boolean, or null)" },
          "timestamp": { "type": "string", "format": "date-time" }
        },
        "required": ["entityType", "entityId", "fieldName", "newValue", "timestamp"]
      },
      "SyncPushRequest": {
        "type": "object",
        "properties": {
          "deviceId": { "type": "string", "example": "device-abc-123" },
          "batchId": { "type": "string", "format": "uuid" },
          "changes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ChangeRecord" },
            "minItems": 1
          }
        },
        "required": ["deviceId", "batchId", "changes"]
      },
      "SyncPushResponse": {
        "type": "object",
        "properties": {
          "batchId": { "type": "string", "format": "uuid" },
          "accepted": { "type": "integer", "example": 2 },
          "conflicts": { "type": "integer", "example": 0 }
        },
        "required": ["batchId", "accepted", "conflicts"]
      },
      "SyncPullRequest": {
        "type": "object",
        "properties": {
          "deviceId": { "type": "string", "example": "device-xyz-456" },
          "lastSyncAt": { "type": "string", "format": "date-time" }
        },
        "required": ["deviceId", "lastSyncAt"]
      },
      "SyncPullResponse": {
        "type": "object",
        "properties": {
          "changes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ChangeRecord" }
          },
          "serverTime": { "type": "string", "format": "date-time" }
        },
        "required": ["changes", "serverTime"]
      }
    }
  },
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
...
    "/api/v1/tags": {
      "post": {
        "tags": ["tags"],
        "summary": "Create a tag",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTagRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tag created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Tag" }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["tags"],
        "summary": "Get all tags",
        "description": "Returns all tags for the authenticated user.",
        "responses": {
          "200": {
            "description": "Array of tags.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Tag" }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tags/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" },
          "description": "Tag UUID"
        }
      ],
      "patch": {
        "tags": ["tags"],
        "summary": "Update a tag",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTagRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag updated.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Tag" }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["tags"],
        "summary": "Delete a tag",
        "description": "Permanently deletes the tag and removes all task associations.",
        "responses": {
          "204": {
            "description": "Tag deleted."
          },
          "400": {
            "description": "Invalid tag ID.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tags/{id}/tasks": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" },
          "description": "Tag UUID"
        }
      ],
      "get": {
        "tags": ["tags"],
        "summary": "Get tasks by tag",
        "description": "Returns all tasks associated with the given tag.",
        "responses": {
          "200": {
            "description": "Array of tasks.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          },
          "400": {
            "description": "Invalid tag ID.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{taskId}/tags/{tagId}": {
      "parameters": [
        {
          "name": "taskId",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" },
          "description": "Task UUID"
        },
        {
          "name": "tagId",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" },
          "description": "Tag UUID"
        }
      ],
      "post": {
        "tags": ["tags"],
        "summary": "Add a tag to a task",
        "description": "Associates the tag with the task. Idempotent.",
        "responses": {
          "204": {
            "description": "Tag added to task."
          },
          "400": {
            "description": "Invalid ID format.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["tags"],
        "summary": "Remove a tag from a task",
        "description": "Removes the tag association from the task.",
        "responses": {
          "204": {
            "description": "Tag removed from task."
          },
          "400": {
            "description": "Invalid ID format.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/push": {
      "post": {
        "tags": ["sync"],
        "summary": "Push local changes to the server",
        "description": "Applies a batch of per-field change records atomically. Returns the count of accepted and conflicted changes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SyncPushRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Push processed.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SyncPushResponse" }
              }
            }
          },
          "400": {
            "description": "Validation error (missing deviceId, batchId, or empty changes).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Push failed due to server error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/pull": {
      "post": {
        "tags": ["sync"],
        "summary": "Pull remote changes from the server",
        "description": "Returns all changes made by other devices after the given lastSyncAt timestamp. Excludes changes from the requesting device.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SyncPullRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Changes retrieved.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SyncPullResponse" }
              }
            }
          },
          "400": {
            "description": "Validation error (missing deviceId).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Pull failed due to server error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    }
  }
}